Correctly encode nonce for API request signatures
Update ABI encoding for nonce from uint256 to uint64 in aster-client.ts, aster-code.ts, miniapp-routes.ts, and telegram-bot.ts to ensure correct signature generation for API requests.

+++ server/aster-client.ts
@@ -302,7 +302,7 @@ async function signV3Params(
   const { AbiCoder, keccak256, getBytes } = await import("ethers");
   const abiCoder = AbiCoder.defaultAbiCoder();
   const encoded = abiCoder.encode(
-    ["string", "address", "address", "uint256"],
+    ["string", "address", "address", "uint64"],
     [queryString, user, signer, BigInt(nonce)],
   );
   const hash = keccak256(encoded);
