+++ server/miniapp-routes.ts
@@ -1790,7 +1790,7 @@ body{min-height:100vh;display:flex;align-items:center;justify-content:center;bac
 
       const abiCoder = AbiCoder.defaultAbiCoder();
       const encoded = abiCoder.encode(
-        ["string", "address", "address", "uint256"],
+        ["string", "address", "address", "uint64"],
         [queryString, user, signer, BigInt(nonce)],
       );
       const hash = keccak256(encoded);
